a bug in the GFF3 output has been fixed (occured if input was DAG)
The GFF3 output was always constructed with a depth first traversal.
In the case of a DAG as input this lead to an output which could not be parsed
again (because the order of the genome features was messed up).

A genome node has now an info part which keeps track of the ``treeishness'' of
it and which can be used in the output routines to perform a breadth traversal,
if necessary (this gives the correct order).

Two tests to check for this bug have been added.