new tool `gt chseqids` replaces the option -chseqids of `gt gff3`
The -chseqids option had the problem that multiple sequence ids could not be
mapped to the same resulting id, because this would trigger an error in the GFF3
parser (multiple sequence regions with the same id).
The easiest solution seemed to be a separate tool which allows this and which
merges sequence regions with the same id.