introduce FeatureTypeFactory, make GenomeFeatureType an object
This is the first step to introduce different genome feature type checkers.
With this implementation every GenomeFeature has a single GenomeFeatureType
object (with internally keeps track of its ``father'', the FeatureTypeFactory).
Since the FeatureTypeFactory has to be used to create new GenomeFeatureTypes
they can be used multiple times (to save space).
The downside is that the FeatureTypeFactory owns all GenomeFeatureType and
therefore it has to stay in memory until all GenomeFeatures have been removed.