gtruby: add GT::Diagram.release_track_selector_func
This method should be called whenever a reference to a GT::Diagram
object with a defined custom track selector function is about to be
dropped (e.g. by expectedly going out of scope). This is a workaround
for a limitation in Ruby-DL allowing only a globally fixed number of
simultaneously defined callbacks.