split ltr/ code into libgenometools and libgtunstable
Due to dependencies in LTRharvest that are not moved into libgenometools.so
yet, the LTRharvest code part of ltr/ remains compiled into libgtunstable.so.
Because the code is needed for the scripting language bindings, the LTRdigest
code (which already largely conforms to the GenomeTools coding style) is then
compiled into to libgenometools.so.