actually handle Lua callback runtime errors
This completes a quick hack introduced earlier which allowed Lua callback
functions to produce node-specific style values. However, error messages
from the callbacks were quietly discarded, making it difficult to debug
custom functions.
This commit fixes this by properly checking the results of protected
calls to Lua for errors and forwarding the messages to the drawing
engine.

This commit removes some functionality from the Style file due to
design issues that are rather solved by reimplementation.
In particular, the type-specific options 'max_num_lines' and
'split_lines' have been removed and are only usable via their
global counterpart in the 'format' section.
This is due to the fact that they do not actually reference types,
but rather tracks, whose caption strings may not be identical to
types.