introduce GtDescBuffer
The current description handling during the parsing of sequence files
copies each description to a separate memory location and stores
a pointer to it in a queue, to be retrieved when using the
GtSequenceBuffer. Afterwards the memory for the descriptions is almost
always freed, making this a malloc/copy/free overhead linear with the number
of sequences. This commit introduces a buffer which reuses memory for
descriptions using parsing, and requires only one small copy operation
whenever the sequence output buffer is empty, instead of at every
sequence.
This comes with a slight interface change as, obviously, ownership
of description string memory now has to remain with the buffer object.