small patches
Assert that the pointer to the memory is not NULL in
GT_CHECKARRAYSPACEMULTI like it is done in GT_CHECKARRAYSPACE. Remove
a superfluous assertion in GT_STOREINARRAY.

Add missing semicolons after GT_CHECKARRAYSPACEMULTI

Wrap all the macros in arraydef.h in do..while blocks

Do not include the human readable representation of errno in the error
messages should malloc, calloc or realloc fail. These functions only
set errno to ENOMEM according to standard anyway, and for
malloc implementations which do not set errno (e.g. the one used in Valgrind)
misleading error messages are printed.

Thanks to Justus Winter.