gtruby: be more economical with scarce callback functions
The Ruby::DL engine only allows for 10 globally allocated callback
functions. This makes it impractical to register a callback for each
possible node type handler in a custom visitor if it is not implemented.
This commit only registers callbacks for node types for which handlers
are present in the GT::CustomVisitor subclass. This can give you 1-2
additional possible custom visitors before the limit kicks in.