bug fix: create transient edges for parent
The transient part_of edges have to be created for the parent in a "has parent"
check (and not for the child). In earlier versions they have been created for
the child and this worked for the standard gene example, because in that case
the mRNA parent was a gene child before. In added test file that is not the case
and therefore it failed erroneously.

This commit also documents how the transient edges are created in more detail.

Resolves #514