Correct error in find_min method of priority_queue.
The method gt_priority_queue_extract_min returns a pointer to the minimal
element. The method gt_priority_queue_find_min however unexpectedly returned
a pointer to the pointer to the minimal element. I have searched through
the code base and as far as I can see, gt_priority_queue_find_min is not
called from anywhere in the code. I expect that that is the reason, why
the error has not been spotted before.