Remove assertion. Bugfix.
The case that this assertion tests can actually happen. It is not an error
when condenseq->ldb_nelems equals 0. It just means that the compression
is bad because no redundancies were found so that the number of links
is zero. To see an example when this happens, execute
gt condenseq compress sequence
and let sequence be the encoded sequence of the sequence found under
http://www.ncbi.nlm.nih.gov/nuccore/NC_000913.3
After removal of the assertion everything is fine. Otherwise we get
an assertion failure.