Fix inclusion of lpeg header
* src/external/lpeg-0.10.2/lua-lpeg.h: New symbolic link.
* src/gtr.c: Include 'lua-lpeg.h' instead.

The Debian source package 'lua-lpeg' synthesizes a header file
'lua-lpeg.h' based on the package name.  This file is generated
because the library no longer ships a header file in more recent
versions.  For details, see the Debian Lua policy:

http://pkg-lua.alioth.debian.org/policy.html#sec-makefileconf

It is unclear to me how you want to handle this problem.