use concurrent communication with hmmscan
We noticed that the current implementation of passing data to and
receiving data from HMMER's hmmscan in the GtLTRdigestPdomVisitor relies
on the pipe buffer to store input and output while a single thread is
handling writing and reading sequentially. This obviously causes
problems once the buffer size is exceeded, with one side of the
communication waiting for the other to read the flushed buffer contents.
This resulted in #970 and probably also #962.
This commit uses separate threads for the read and write sides of the
communication.